# docs/swagger.yaml
definitions:
...
  controllers.crew:
    properties:
      id:
        example: 24310
        type: integer
      name:
        example: Mitchell Amundsen
        type: string
      profileUrl:
        example: https://image.tmdb.org/t/p/original/zK4o3lfe6chC2qDHhxpuO1RJl2X.jpg
        type: string
      role:
        example: Director
        type: string
    type: object
  controllers.errorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
  controllers.genre:
    properties:
      id:
        example: 35
        type: integer
      name:
        example: Comédie
        type: string
    type: object
...
  controllers.movieResponse:
    properties:
      actors:
        items:
          $ref: '#/definitions/controllers.person'
        type: array
      backdropUrl:
        example: https://image.tmdb.org/t/p/original/e7FzphKs5gzoghDotAEp2FeP46u.jpg
        type: string
      crew:
        items:
          $ref: '#/definitions/controllers.crew'
        type: array
      genres:
        items:
          $ref: '#/definitions/controllers.genre'
        type: array
      id:
        example: 649609
        type: integer
      overview:
        example: "Le mal ne saurait survivre une éternité sans un petit coup de pouce.\r
          Dans cette version moderne du mythe de Dracula..."
        type: string
      posterUrl:
        example: https://image.tmdb.org/t/p/original/lm3y4RNPu4aRDePsX5CkB9ndEdQ.jpg
        type: string
      present:
        example: true
        type: boolean
      releaseDate:
        example: "2023-04-07"
        type: string
      studios:
        items:
          $ref: '#/definitions/controllers.studio'
        type: array
      title:
        example: Renfield
        type: string
      voteAverage:
        example: 7.252
        type: number
      voteCount:
        example: 278
        type: integer
    type: object
  controllers.movieResults:
    properties:
      results:
        items:
          $ref: '#/definitions/controllers.movieResponse'
        type: array
      totalPage:
        example: 71
        type: integer
      totalResult:
        example: 1412
        type: integer
    type: object
  controllers.person:
    properties:
      character:
        example: R.M. Renfield
        type: string
      id:
        example: 3292
        type: integer
      name:
        example: Nicholas Hoult
        type: string
      profileUrl:
        example: https://image.tmdb.org/t/p/original/rbyi6sOw0dGV3wJzKXDopm2h0NO.jpg
        type: string
    type: object
  controllers.studio:
    properties:
      id:
        example: 33
        type: integer
      logoUrl:
        example: https://image.tmdb.org/t/p/original/8lvHyhjr8oUKOOy2dKXoALWKdp0.png
        type: string
      name:
        example: Universal Pictures
        type: string
    type: object
...
  controllers.tvEpisodeResponse:
    properties:
      airDate:
        example: "2023-03-24"
        type: string
      episodeNumber:
        example: 12
        type: integer
      id:
        example: 4137463
        type: integer
      name:
        example: Le plus puissant sorcier du monde révèle Akasha
        type: string
      overview:
        example: Ray et ses amis volent au secours de Rebecca, qui montre des signes...
        type: string
      posterUrl:
        example: https://image.tmdb.org/t/p/original/uVqsuh8qrNX8tkQDpDF7nDZdg0w.jpg
        type: string
      present:
        example: true
        type: boolean
      seasonNumber:
        example: 1
        type: integer
      tvShowId:
        example: 200777
        type: integer
    type: object
  controllers.tvShowResponse:
    properties:
      actors:
        items:
          $ref: '#/definitions/controllers.person'
        type: array
      backdropUrl:
        example: https://image.tmdb.org/t/p/original/oL459mgvcnc3jL90K7zkfvXQu0.jpg
        type: string
      crew:
        items:
          $ref: '#/definitions/controllers.crew'
        type: array
      genres:
        items:
          $ref: '#/definitions/controllers.genre'
        type: array
      id:
        example: 200777
        type: integer
      networks:
        items:
          $ref: '#/definitions/controllers.studio'
        type: array
      nextEpisode:
        $ref: '#/definitions/controllers.tvEpisodeResponse'
      overview:
        example: Ray White est un jeune homme venant d'entrer dans la populaire académie
          de magie Arnold. En tant que...
        type: string
      posterUrl:
        example: https://image.tmdb.org/t/p/original/aiJd0oGkBhf98uEH3F3yC7O48vr.jpg
        type: string
      present:
        example: true
        type: boolean
      releaseDate:
        example: "2023-01-06"
        type: string
      seasonsCount:
        example: 1
        type: integer
      status:
        example: Ended
        type: string
      title:
        example: The Iceblade Sorcerer Shall Rule the World
        type: string
      voteAverage:
        example: 6.7
        type: number
      voteCount:
        example: 11
        type: integer
    type: object
  controllers.tvShowResults:
    properties:
      results:
        items:
          $ref: '#/definitions/controllers.tvShowResponse'
        type: array
      totalPage:
        example: 71
        type: integer
      totalResult:
        example: 1412
        type: integer
    type: object
info:
  contact: {}
  description: |-
    This is the API for the Media Service application
    This help to give info about the media files and metadata
    This also help to manage the media files for admins
  title: Media Service API
paths:
...
  /discover/movie/actor:
    get:
      description: Get movies by actor
      parameters:
      - description: Actor id
        in: query
        name: actor
        required: true
        type: integer
      - description: Page number
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.movieResults'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.errorResponse'
      summary: Get movies by actor
      tags:
      - Discover
      - Movie
  /discover/movie/director:
    get:
      description: Get movies by director
      parameters:
      - description: Director id
        in: query
        name: director
        required: true
        type: integer
      - description: Page number
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.movieResults'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.errorResponse'
      summary: Get movies by director
      tags:
      - Discover
      - Movie
  /discover/movie/genre:
    get:
      description: Get movies by genre
      parameters:
      - description: Genre id
        in: query
        name: genre
        required: true
        type: integer
      - description: Page number
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.movieResults'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.errorResponse'
      summary: Get movies by genre
      tags:
      - Discover
      - Movie
  /discover/movie/popular:
    get:
      description: Get popular movies
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Only available movies
        in: query
        name: available
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.movieResults'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.errorResponse'
      summary: Get popular movies
      tags:
      - Discover
      - Movie
  /discover/movie/recent:
    get:
      description: Get recent movies
      parameters:
      - description: Only available movies
        in: query
        name: available
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/controllers.movieResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.errorResponse'
      summary: Get recent movies
      tags:
      - Discover
      - Movie
  /discover/movie/recommendations/{movie}:
    get:
      description: Get movie's recommendations
      parameters:
      - description: Movie id
        in: path
        name: movie
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/controllers.movieResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.errorResponse'
      summary: Get movie's recommendations
      tags:
      - Discover
      - Movie
  /discover/movie/search:
    get:
      description: Search movies by query
      parameters:
      - description: Search query
        in: query
        name: query
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Only available movies
        in: query
        name: available
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.movieResults'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.errorResponse'
      summary: Search movies
      tags:
      - Discover
      - Movie
  /discover/movie/studio:
    get:
      description: Get movies by studio
      parameters:
      - description: Studio id
        in: query
        name: studio
        required: true
        type: integer
      - description: Page number
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.movieResults'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.errorResponse'
      summary: Get movies by studio
      tags:
      - Discover
      - Movie
  /discover/tv/actor:
    get:
      description: Get tv shows by actor
      parameters:
      - description: Actor id
        in: query
        name: actor
        required: true
        type: integer
      - description: Page number
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.tvShowResults'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.errorResponse'
      summary: Get tv shows by actor
      tags:
      - Discover
      - TvShow
  /discover/tv/genre:
    get:
      description: Get tv shows by genre
      parameters:
      - description: Genre id
        in: query
        name: genre
        required: true
        type: integer
      - description: Page number
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.tvShowResults'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.errorResponse'
      summary: Get tv shows by genre
      tags:
      - Discover
      - TvShow
  /discover/tv/network:
    get:
      description: Get tv shows by network
      parameters:
      - description: Network id
        in: query
        name: network
        required: true
        type: integer
      - description: Page number
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.tvShowResults'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.errorResponse'
      summary: Get tv shows by network
      tags:
      - Discover
      - TvShow
  /discover/tv/popular:
    get:
      description: Get popular tv shows
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Only available tv shows
        in: query
        name: available
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.tvShowResults'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.errorResponse'
      summary: Get popular tv shows
      tags:
      - Discover
      - TvShow
  /discover/tv/recent:
    get:
      description: Get recent tv shows
      parameters:
      - description: Only available tv shows
        in: query
        name: available
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/controllers.tvShowResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.errorResponse'
      summary: Get recent tv shows
      tags:
      - Discover
      - TvShow
  /discover/tv/recommendations/{tv}:
    get:
      description: Get tv show's recommendations
      parameters:
      - description: Show id
        in: path
        name: show
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/controllers.tvShowResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.errorResponse'
      summary: Get tv show's recommendations
      tags:
      - Discover
      - TvShow
  /discover/tv/search:
    get:
      description: Search tv shows by query
      parameters:
      - description: Search query
        in: query
        name: query
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Only available tv shows
        in: query
        name: available
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.tvShowResults'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controllers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controllers.errorResponse'
      summary: Search tv shows
      tags:
      - Discover
      - TvShow
...
swagger: "2.0"